changed migration & up/down scripts
