--- commands/create_database_json.js
@@ -1,2 +1,0 @@
-var installation = require('./installation.js');
-installation.createDatabaseJson();
