--- commands/down.sh
@@ -1,4 +1,0 @@
-SCRIPT=$(readlink -f "$0")
-SCRIPTPATH=$(dirname "$SCRIPT")
-kill $(cat $SCRIPTPATH/pid)
-rm $SCRIPTPATH/pid
