+++ install/install.sh
@@ -1,4 +1,4 @@
 SCRIPT=$(readlink -f "$0")
 SCRIPTPATH=$(dirname "$SCRIPT")
-nodejs $SCRIPTPATH/create_database_json.js
+nodejs $SCRIPTPATH/prepare_migration.js
 nodejs $SCRIPTPATH/../node_modules/.bin/db-migrate up
