+++ package.json
@@ -14,7 +14,8 @@
     "cookie-parser": "~1.3",
     "morgan": "~1.6",
     "db-migrate": "~0.9",
-    "node-sass": "~3.2"
+    "node-sass": "~3.2",
+    "forever": "~0.15"
   },
   "devDependencies": {
     "mocha": "~2.2",
